Line rendering API
You can either draw a line from 2 points, draw connected lines with an array of points with a count, or just passing directly the points. It can be everything for that last one, not only Vector2f, and should compile if the passing argument can be implicitly converted to a Vector2f. But for preference, use Vector2f